Refactor: alarm directory
