+++ green/src/main/java/com/greengrim/green/common/fcm/FcmMessageBuilder.java
@@ -1,7 +1,7 @@
 package com.greengrim.green.common.fcm;
 
 import com.google.firebase.messaging.Message;
-import com.greengrim.green.core.alarm.AlarmType;
+import com.greengrim.green.core.alarm.entity.AlarmType;
 import com.greengrim.green.core.chat.ChatMessage;
 import com.greengrim.green.core.member.Member;
 import org.springframework.stereotype.Component;
