+++ green/src/main/java/com/greengrim/green/core/alarm/service/GetAlarmService.java
@@ -1,7 +1,9 @@
-package com.greengrim.green.core.alarm;
+package com.greengrim.green.core.alarm.service;
 
 import com.greengrim.green.common.entity.dto.PageResponseDto;
 import com.greengrim.green.core.alarm.dto.AlarmResponseDto.AlarmInfo;
+import com.greengrim.green.core.alarm.entity.Alarm;
+import com.greengrim.green.core.alarm.repository.AlarmRepository;
 import com.greengrim.green.core.member.Member;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
@@ -17,27 +19,9 @@ import java.util.List;
 
 @Service
 @RequiredArgsConstructor
-public class AlarmService {
+public class GetAlarmService {
 
     private final AlarmRepository alarmRepository;
-
-    @Transactional
-    public Alarm register(Member member, AlarmType alarmType, Long resourceId,
-                          String imgUrl, String variableContent, Long interactedMemberId) {
-        if(interactedMemberId == null) {
-            interactedMemberId = -1L;
-        }
-        Alarm alarm = Alarm.builder()
-                .member(member)
-                .type(alarmType)
-                .resourceId(resourceId)
-                .imgUrl(imgUrl)
-                .variableContent(variableContent)
-                .interactedMemberId(interactedMemberId)
-                .build();
-        alarmRepository.save(alarm);
-        return alarm;
-    }
 
     @Transactional
     public PageResponseDto<List<AlarmInfo>> getAlarms(Member member, int page, int size) {
